+++ Cargo.toml
@@ -60,11 +60,13 @@ ethereum-tx-sign = "3.0.4"
 anomaly = "0.2"
 secp256k1 = {version = "0.19", features = ["recovery"] }
 
+jsonrpc-core = "17.1.0"
+jsonrpc-core-client = "17.1.0"
+jsonrpc-ws-server = "17.1.0"
+jsonrpc-derive = "17.1.0"
+
 [dev-dependencies]
-abscissa_core = { version = "=0.6.0-pre.1", features = ["testing"] }
-byteorder = "1"
 rand = "0.7"
-smol = "1.2.5"
 
 [features]
 softsign = []
