+++ src/commands/ether.rs
@@ -4,10 +4,12 @@
 //mod keygen;
 
 mod import;
+mod start;
 
 //use self::{import::ImportCommand, keygen::KeygenCommand};
 use abscissa_core::{Command, Help, Options, Runnable};
 use crate::commands::ether::import::ImportCommand;
+use crate::commands::ether::start::StartCommand;
 
 /// 'ether' command: provides subcommands for local Ethereum signer
 #[derive(Command, Debug, Options, Runnable)]
@@ -16,5 +18,6 @@ pub enum EtherCommand {
     Help(Help<Self>),
     #[options(help = "import Ethereum key")]
     Import(ImportCommand),
-
+    #[options(help = "Start JSON-RPC Ethereum signing service")]
+    Start(StartCommand),
 }
