--- src/commands/ether/sign.rs
@@ -1,8 +1,0 @@
-//! `tmkms ether sign` command
-
-use abscissa_core::{Command, Options, Runnable};
-
-#[derive(Command, Debug, Default, Options)]
-pub struct SignCommand {
-
-}
